Fix how discord IDs are gotten.
