+++ src/backend/controllers/user/UserController.ts
@@ -5,6 +5,7 @@ import {UserDTO} from "mnemoshared/dist/src/dto/model/UserDTO";
 import {DataDTO} from "mnemoshared/dist/src/dto/model/DataDTO";
 import {WorldDTO} from "mnemoshared/dist/src/dto/model/WorldDTO";
 import {CharacterDTO} from "mnemoshared/dist/src/dto/model/CharacterDTO";
+import {Collection} from "discord.js";
 
 @injectable()
 export class UserController extends API<UserDTO> {
@@ -79,6 +80,42 @@ export class UserController extends API<UserDTO> {
         return this.put(`/users/${discordId}`, config).then((res) => {
             // @ts-ignore
             return res.data.data;
+        }).catch((err: Error) => {
+            console.log("Caught error.");
+            console.error(err);
+            return null;
+        });
+    }
+
+    /**
+     * Gets all the discord IDs related to this character.
+     * @param characterIds
+     * @param worldId
+     */
+    public async getDiscordId(characterIds: string[], worldId: string): Promise<Collection<string, string>> {
+        let config = APIConfig.GET();
+        config.params = {
+            character_ids: characterIds,
+            world_id: worldId
+        };
+
+        return this.get(`/discordIds`, config).then((res) => {
+            // @ts-ignore
+            if (!res || !res.data || !res.data.data) {
+                return null;
+            }
+
+            // @ts-ignore
+            let ids: string[] = res.data.data;
+
+            let input = new Collection<string, string>();
+            let discordId: string, i;
+            for (i = 0; i < ids.length; i++) {
+                discordId = ids[i];
+                input.set(discordId, discordId);
+            }
+
+            return input;
         }).catch((err: Error) => {
             console.log("Caught error.");
             console.error(err);
